third_party/blink/web_tests/platform/generic/fast/forms/ValidityState-tooShort-input-expected.txt

Tests for tooShort flag with <input> elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

No minlength and no value
PASS input.validity.tooShort is false


Non-dirty value
PASS input.value.length is 2
PASS input.validity.tooShort is false
PASS input.value.length is 1
PASS input.validity.tooShort is false


Dirty value and longer than minLength
PASS input.value.length is 1
PASS input.validity.tooShort is true
PASS input.value.length is 0
PASS input.validity.tooShort is false
PASS input.value.length is 2
PASS input.validity.tooShort is true
PASS input.value.length is 3
PASS input.validity.tooShort is false


Sets a value via DOM property
PASS input.validity.tooShort is false


Disabling makes the control valid
PASS input.validity.tooShort is true
PASS input.disabled = true; input.validity.tooShort is true
PASS input.disabled = false; input.validity.tooShort is true


Change the type with a too long value
PASS input.validity.tooShort is true
PASS input.type = "number"; input.validity.tooShort is false


Grapheme length is shorter than minLength though character length is greater
PASS input.validity.tooShort is false
PASS input.value.length is 3
PASS input.validity.tooShort is false
PASS successfullyParsed is true

TEST COMPLETE
